requirements.txt
python-dotenv
discord.py
discord
py-mon
colorama
watchdog
requests